Cleanup listener stuff
