+++ Insights-API/src/main/java/dev/frankheijden/insights/api/InsightsMain.java
@@ -13,9 +13,8 @@ import dev.frankheijden.insights.api.config.Limits;
 import dev.frankheijden.insights.api.config.Messages;
 import dev.frankheijden.insights.api.config.Notifications;
 import dev.frankheijden.insights.api.config.Settings;
+import dev.frankheijden.insights.api.listeners.manager.InsightsListenerManager;
 import dev.frankheijden.insights.api.metrics.MetricsManager;
-import java.lang.reflect.Method;
-import java.util.Map;
 
 public interface InsightsMain {
 
@@ -55,7 +54,6 @@ public interface InsightsMain {
 
     ScanHistory getScanHistory();
 
-    Map<String, Method> getAllowedDisableMethods();
+    InsightsListenerManager getListenerManager();
 
-    Map<String, Method> getAllowedPriorityOverrideMethods();
 }
